requirements.txt
django>=4.2,<5.0
djangorestframework==3.16.1
psycopg2-binary==2.9.11
python-decouple==3.8
pillow==12.1.0
dj-database-url==3.1.0
python-dotenv==1.2.1
requests==2.32.5
celery==5.6.2
redis==7.1.0
django-allauth==65.14.0
dj-rest-auth==7.0.2
djangorestframework-simplejwt==5.5.1
drf-yasg==1.21.14
Faker==22.5.1
torch>=2.0.0
torchvision>=0.15.0
faiss-cpu>=1.7.0
numpy>=1.24.0